l1: convert events_from_other_contracts test into integration test
It requires anvil, so should not be run as a unit test
