+++ crates/papyrus_base_layer/src/anvil_base_layer.rs
@@ -7,6 +7,7 @@ use alloy::rpc::types::TransactionReceipt;
 use async_trait::async_trait;
 use colored::*;
 use starknet_api::block::BlockHashAndNumber;
+use starknet_api::hash::StarkHash;
 use starknet_api::transaction::L1HandlerTransaction;
 use url::Url;
 
@@ -33,6 +34,8 @@ pub struct AnvilBaseLayer {
 }
 
 impl AnvilBaseLayer {
+    pub const DEFAULT_ANVIL_L1_ACCOUNT_ADDRESS: StarkHash =
+        StarkHash::from_hex_unchecked("0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266");
     const DEFAULT_ANVIL_PORT: u16 = 8545;
     const DEFAULT_ANVIL_L1_DEPLOYED_ADDRESS: &str = "0x5fbdb2315678afecb367f032d93f642f64180aa3";
