+++ crates/papyrus_base_layer/src/base_layer_test.rs
@@ -3,22 +3,16 @@ use alloy::primitives::B256;
 use alloy::providers::mock::Asserter;
 use alloy::providers::{Provider, ProviderBuilder};
 use alloy::rpc::types::{Block, BlockTransactions, Header as AlloyRpcHeader};
-use assert_matches::assert_matches;
 use pretty_assertions::assert_eq;
 use starknet_api::block::{BlockHash, BlockHashAndNumber, BlockNumber};
-use starknet_api::core::EntryPointSelector;
-use starknet_api::transaction::L1HandlerTransaction;
-use starknet_api::{calldata, contract_address, felt};
+use starknet_api::felt;
 
-use crate::anvil_base_layer::{send_message_to_l2, AnvilBaseLayer};
-use crate::constants::{EventIdentifier, LOG_MESSAGE_TO_L2_EVENT_IDENTIFIER};
 use crate::ethereum_base_layer_contract::{
     EthereumBaseLayerConfig,
     EthereumBaseLayerContract,
     Starknet,
 };
-use crate::test_utils::DEFAULT_ANVIL_L1_ACCOUNT_ADDRESS;
-use crate::{BaseLayerContract, L1Event};
+use crate::BaseLayerContract;
 
 // TODO(Gilad): Use everywhere instead of relying on the confusing `#[ignore]` api to mark slow
 // tests.
@@ -111,61 +105,4 @@ async fn get_gas_price_and_timestamps() {
     // Roughly e ** (BLOB_GAS / eip4844::BLOB_GASPRICE_UPDATE_FRACTION)
     let expected_original_blob_calc = 19;
     assert_eq!(header.blob_fee, expected_original_blob_calc);
-}
-
-// Ensure that the base layer instance filters out events from other deployments of the core
-// contract.
-#[tokio::test]
-async fn events_from_other_contract() {
-    if !in_ci() {
-        return;
-    }
-    const EVENT_IDENTIFIERS: &[EventIdentifier] = &[LOG_MESSAGE_TO_L2_EVENT_IDENTIFIER];
-
-    let anvil_base_layer = AnvilBaseLayer::new().await;
-    // Anvil base layer already auto-deployed a starknet contract.
-    let this_contract = &anvil_base_layer.ethereum_base_layer.contract;
-
-    // Setup.
-
-    // Deploy another instance of the contract to the same anvil instance.
-    let other_contract = Starknet::deploy(this_contract.provider().clone()).await.unwrap();
-    assert_ne!(
-        this_contract.address(),
-        other_contract.address(),
-        "The two contracts should be different."
-    );
-
-    let this_l1_handler = L1HandlerTransaction {
-        contract_address: contract_address!("0x12"),
-        entry_point_selector: EntryPointSelector(felt!("0x34")),
-        calldata: calldata!(DEFAULT_ANVIL_L1_ACCOUNT_ADDRESS, felt!("0x1"), felt!("0x2")),
-        ..Default::default()
-    };
-    let this_receipt = send_message_to_l2(this_contract, &this_l1_handler.clone()).await;
-    assert!(this_receipt.status());
-    let this_block_number = this_receipt.block_number.unwrap();
-
-    let other_l1_handler = L1HandlerTransaction {
-        contract_address: contract_address!("0x56"),
-        entry_point_selector: EntryPointSelector(felt!("0x78")),
-        calldata: calldata!(DEFAULT_ANVIL_L1_ACCOUNT_ADDRESS, felt!("0x1"), felt!("0x2")),
-        ..Default::default()
-    };
-    let other_receipt = send_message_to_l2(&other_contract, &other_l1_handler.clone()).await;
-    assert!(other_receipt.status());
-    let other_block_number = other_receipt.block_number.unwrap();
-
-    let min_block_number = this_block_number.min(other_block_number).saturating_sub(1);
-    let max_block_number = this_block_number.max(other_block_number).saturating_add(1);
-
-    // Test the events.
-    let mut events = anvil_base_layer
-        .ethereum_base_layer
-        .events(min_block_number..=max_block_number, EVENT_IDENTIFIERS)
-        .await
-        .unwrap();
-
-    assert_eq!(events.len(), 1, "Expected only events from this contract.");
-    assert_matches!(events.remove(0), L1Event::LogMessageToL2 { tx, .. } if tx == this_l1_handler);
 }
